Add clustering_utils
